Confirmation suppression

+++ server/services/appointmentService.js
@@ -89,6 +89,7 @@ export async function createAppointment(user, body) {
  * Modification d’un rendez-vous
  */
 export async function updateAppointment(user, body) {
+    //mettre ici comme quoi on recupère only ou all pour savoir ce qu'on a choisi
     const {
         id,
         idRegle,
@@ -102,7 +103,14 @@ export async function updateAppointment(user, body) {
         frequenceId,
         fin_rec,
         date_fin_rec,
+        modifRec,
+        day,
+        month,
+        year,
     } = body;
+ 
+    const jour = parseInt(day) +1; 
+    const dateException = new Date(year, month, jour);
 
     const appointment = await AppointmentModel.findById(id).populate("agenda");
     if (!appointment) throw new Error("Rendez-vous introuvable");
@@ -122,46 +130,105 @@ export async function updateAppointment(user, body) {
     const dateFin = buildDate(date_fin, heure_fin);
 
     let updatedRecRule = null;
+    let appointmentNew = null;
+    let updated = null;
 
     if (recurrence === "on") {
-        let dateFinRec = fin_rec === "never" ? null : new TZDate(date_fin_rec);
+        // Si on modifie que un seul rdv
+        if(modifRec === "only"){
+            appointmentNew = new AppointmentModel({
+                agenda: newAgenda._id,
+                nom,
+                date_Debut: dateDebut,
+                date_Fin: dateFin,
+                recurrenceRule: null,
+                modifRecurrence: true,
+                createur: user._id,
+            });
+            await appointmentNew.save();
+        }else{
+            let dateFinRec = fin_rec === "never" ? null : new TZDate(date_fin_rec);
 
-        // Si une règle existe -> update
-        if (appointment.recurrenceRule) {
-            updatedRecRule = await RegleOccurrenceModel.findByIdAndUpdate(
-                idRegle,
-                {
+            // Si une règle existe -> update
+            if (appointment.recurrenceRule) {
+                updatedRecRule = await RegleOccurrenceModel.findByIdAndUpdate(
+                    idRegle,
+                    {
+                        frequence: frequenceId,
+                        date_fin: dateFinRec,
+                    },
+                    { new: true }
+                );
+            } else {
+                // Sinon créer une nouvelle règle
+                updatedRecRule = new RegleOccurrenceModel({
                     frequence: frequenceId,
                     date_fin: dateFinRec,
-                },
-                { new: true }
-            );
-        } else {
-            // Sinon créer une nouvelle règle
-            updatedRecRule = new RegleOccurrenceModel({
-                frequence: frequenceId,
-                date_fin: dateFinRec,
-            });
-            await updatedRecRule.save();
+                });
+                await updatedRecRule.save();
+            }
         }
     } else {
         // Suppression règle
         updatedRecRule = null;
+        /* appointmentNew = new AppointmentModel({
+                agenda: newAgenda._id,
+                nom,
+                date_Debut: dateDebut,
+                date_Fin: dateFin,
+                recurrenceRule: null,
+                createur: user._id,
+            });
+            await appointmentNew.save(); */
     }
 
-    // Mise à jour RDV
-    const updated = await AppointmentModel.findByIdAndUpdate(
-        id,
-        {
-            agenda: newAgenda._id,
-            nom,
-            date_Debut: dateDebut,
-            date_Fin: dateFin,
-            recurrenceRule: updatedRecRule?._id || null,
-        },
-        { new: true }
-    );
+    if(modifRec === "only" && recurrence === "on"){
+        //Si on modifie que 1 seul rdv et qu'il y a encore la récurrence d'activer, il fait encore parti de la rec et donc on l'ajoute comme lié à la rec encore
+        updated = await AppointmentModel.findByIdAndUpdate(
+            id,
+            {
+                $addToSet: { 
+                    exception: appointmentNew._id,
+                    exceptionDate: appointmentNew.date_Debut 
+                },
+            },
+            { new: true }
+        );
+    }else if (modifRec != "only"){ //passe pour all et off ou on
+        // si il est sur all, on supprime le lien avec la récurrence si c'est undefined et on garde le lien si c'est on
+        // Mise à jour RDV
+        updated = await AppointmentModel.findByIdAndUpdate(
+            id,
+            {
+                agenda: newAgenda._id,
+                nom: nom,
+                date_Debut: dateDebut,
+                date_Fin: dateFin,
+                recurrenceRule: updatedRecRule?._id || null,
+            },
+            { new: true }
+        );
+    }else if(modifRec === "only"){ //isoler un rdv de la récurrence pour l'enlever de la récurrence
+        appointmentNew = new AppointmentModel({
+                agenda: newAgenda._id,
+                nom,
+                date_Debut: dateDebut,
+                date_Fin: dateFin,
+                recurrenceRule: null,
+                createur: user._id,
+            });
+        await appointmentNew.save();
 
+        updated = await AppointmentModel.findByIdAndUpdate(
+            id,
+            {
+                $addToSet: { 
+                    exceptionDate: appointmentNew.date_Debut 
+                },
+            },
+            { new: true }
+        );
+    }
     await creerNotification(user, updated, user, newAgenda, 2);
 
     return updated;
@@ -173,7 +240,10 @@ export async function updateAppointment(user, body) {
  * Suppression d’un rendez-vous
  */
 export async function deleteAppointment(user, body) {
-    const { id } = body;
+    const { id, agendas, modifRecSup } = body;
+
+    //il me faut le champs only ou all qui permet de savoir si je modifie que une unique occurrence ou si je modifie tout
+    //si on supprime que une seule occurence, il faut que je récupère la date et que je 
 
     const appointment = await AppointmentModel.findById(id);
     if (!appointment) throw new Error("Rendez-vous introuvable");
@@ -187,11 +257,35 @@ export async function deleteAppointment(user, body) {
         throw new Error("deleteAppointment : Vous n'avez pas les droits pour supprimer ce rendez-vous.");
     }
 
+    if(appointment.recurrenceRule){
+        const recId = appointment.recurrenceRule;
+        if(modifRecSup === 'all'){
+
+            if(appointment.exception && appointment.exception.length >0){
+                const exceptionIds = appointment.exception.map(exc => exc._id);
+                await AppointmentModel.deleteMany({ _id: { $in: exceptionIds } });
+            }
+            await RegleOccurrenceModel.findByIdAndDelete(recId);
+            await AppointmentModel.findByIdAndDelete(id);
+
+        }else if (modifRecSup === 'only'){
+
+            const dateException = appointment.date_Debut;
+            await AppointmentModel.findByIdAndUpdate(
+            id,
+            {
+                $addToSet: { exceptionDate: dateException },
+            },
+            { new: true }
+        );
+        }     
+    }else{
+        await AppointmentModel.findByIdAndDelete(id);
+    }
+
     await creerNotification(agenda.user, appointment, user, agenda, 3);
 
     await supprimerNotification(id);
-
-    await AppointmentModel.findByIdAndDelete(id);
 
     return true;
 }
@@ -211,7 +305,8 @@ export function buildAppointmentFormData({
 
     const name = appointment?.nom || '';
     const rdvId = appointment?._id?.toString() || null;
-    const recurrence = !!appointment?.recurrenceRule;
+    const recurrence = appointment?.recurrenceRule;
+    const modifRecurrence = appointment?.modifRecurrence;
     let frequenceRegle = '';
     let regleId = null;
     let date_fin_ap = null;
@@ -219,13 +314,18 @@ export function buildAppointmentFormData({
     let requestedDate = parseDate(day, month, year);
     let requestedDateEnd = new TZDate(requestedDate);
     
-    let dateDeb = new TZDate(requestedDate).setHours(beginningHour);
+    let dateDeb = new TZDate(requestedDate);
+    dateDeb.setHours(beginningHour);
+    
     let dateFin = new TZDate(requestedDate).setHours(beginningHour + 1);
     if(appointment) {
-        dateDeb = new TZDate(appointment.date_Debut);
-        requestedDate = formatDate(dateDeb);
-        dateFin = new TZDate(appointment.date_Fin);
-        requestedDateEnd = formatDate(dateFin);
+        if (!recurrence)
+        {
+            dateDeb = new TZDate(appointment.date_Debut);
+            requestedDate = formatDate(dateDeb);
+            dateFin = new TZDate(appointment.date_Fin);
+            requestedDateEnd = formatDate(dateFin);
+        }
     }
 
     let formDate = formatDate(dateDeb);
@@ -250,6 +350,7 @@ export function buildAppointmentFormData({
         name,
         rdvId,
         recurrence,
+        modifRecurrence,
         frequenceRegle,
         regleId,
         date_fin_ap,
